requirements.txt
pyrtlsdr==0.3.0
numpy==1.26.4
scipy==1.13.0
flask==3.0.3
flask-socketio==5.3.6
flask-cors==4.0.1
eventlet==0.36.1
matplotlib==3.9.0
scikit-learn==1.5.0
joblib==1.4.2
pyaudio==0.2.14
